+++ backend/controllers/CreateFeatureFlag.js
@@ -1,0 +1,16 @@
+const FeatureFlagModel = require('../models/FeatureFlag');
+
+/* Create Feature Flag */
+const createFeatureFlag = async (req, res) => {
+    try{
+        if(req.body !== undefined){
+            const feature = new FeatureFlagModel(req.body);
+            feature.save(); 
+        }
+    }
+    catch(err){
+        return `There is an error: ${err}`
+    }
+}
+
+module.exports = createFeatureFlag;
