+++ backend/routes/featureFlag.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const createFeatureFlag = require('../controllers/createFeatureFlag');
+const router = express.Router();
+
+/* Create Feature Flag */
+router.post('/', createFeatureFlag);
+
+module.exports = router;
