+++ backend/routes/fetchFeatureFlags.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const fetchFeatureFlags = require('../controllers/fetchFeatureFlags');
+const router = express.Router();
+
+/* GET Feature Flags */
+router.get('/', fetchFeatureFlags);
+
+module.exports = router;
